+++ vandal/objects/montecarlo.py
@@ -1,5 +1,8 @@
 # coloring.
-from colorama import Fore, init 
+from colorama import (
+    Fore,
+    init,
+) 
 
 init()